@@ -1,8 +1,8 @@
 package com.udemy.spring.springselenium.entity;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
 import java.sql.Date;
 
 @Entity
